VAIO: Visual Basic successfully working.
Configurations for HeadGirl and Admins has been set up alongside a new set up including the Profiles for both posts. A final votes page is still required and will be built as the next stage. Likewise, there will be requirements for poll viewing for the authorised members of staff.

No date has been set for certain years to vote yet, this may be moved as a last minute addition.